moved logic of worktimeAndBreaktimeOnDate from ShiftForm into utils/shift.js
